requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.1
numpy==1.26.4
sentence-transformers==2.5.1
huggingface-hub==0.21.4
PyGithub==2.2.0
python-dotenv==1.0.1
openai==1.13.3
